+++ lib/features/auth/domain/providers/auth_service_provider.dart
@@ -1,5 +1,5 @@
 import 'package:gym/features/auth/auth.dart';
-import 'package:gym/features/auth/data/data.dart';
+import 'package:gym/features/auth/impl/impl.dart';
 import 'package:riverpod/riverpod.dart';
 
 final authServiceProvider = Provider<AuthService>(
