+++ lib/features/profile/domain/domain.dart
@@ -1,0 +1,5 @@
+export 'services/services.dart';
+export 'models/models.dart';
+export 'repositories/repositories.dart';
+export 'providers/providers.dart';
+export 'data_sources/data_sources.dart';
