+++ lib/features/profile/profile.dart
@@ -1,0 +1,2 @@
+export 'domain/domain.dart';
+export 'presentation/presentation.dart';
