+++ lib/features/themes/domain/providers/system_brightness_service_provider.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:gym/features/themes/data/data.dart';
+import 'package:gym/features/themes/impl/impl.dart';
 import 'package:gym/features/themes/domain/domain.dart';
 import 'package:riverpod/riverpod.dart';
